Move helpers in root helpers.py into specific modules
For the life of me, I can't figure out a logical place to put get_full_path other than the root helpers.py so the file lives another day :/
