+++ algorithms/helpers.py
@@ -1,0 +1,22 @@
+import heapq
+from collections import OrderedDict
+from math import inf
+from typing import Dict
+
+
+def top_n(value_dict: dict, n: int) -> Dict[str, float]:
+    """Helper function to find the n highest-value keys in a dictionary.
+    Runs in O(n+k) time for a dictionary with k entries."""
+    if n is None or n == inf:
+        return value_dict
+    # Have to reformat the dict like this for heapq to cooperate.
+    collection = [(value, key) for key, value in value_dict.items()]
+    heapq.heapify(collection)
+    top_n_items = OrderedDict()
+    for nth_largest in heapq.nlargest(min(n, len(collection)), collection):
+        top_n_items[nth_largest[1]] = nth_largest[0]  # Reconstruct the dict
+    return top_n_items
+
+
+def format_reporter(volume, reporter, page):
+    return f"{volume} {reporter} {page}"
