+++ extraction/citation_extractor.py
@@ -1,6 +1,6 @@
 from typing import Iterable, List, cast
 from db.peewee.models import Opinion, Cluster
-from helpers import format_reporter
+from algorithms.helpers import format_reporter
 import eyecite
 from eyecite.models import (Resource as EyeciteResource, CitationBase, CaseCitation)
